Entorno virtual/requirements.txt
Django==5.0.3
psycopg2==2.9.9
djangorestframework==3.15.0
